clean local test
