--- tests/local_tests/minimax_h3/__init__.py
@@ -1,1 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
