+++ tests/local_tests/transformers/test_minimax_h3_transformer_parity.py
@@ -1,135 +1,87 @@
 # SPDX-License-Identifier: Apache-2.0
-"""CPU synthetic parity for the native MiniMax H3 Transformer."""
+"""Production-loader parity for both MiniMax H3 DiT partitions.
+
+This gated CUDA test loads the published ``transformer/`` and
+``transformer_ref/`` checkpoints through FastVideo's production
+``TransformerLoader`` and compares both output heads against the pinned
+Diffusers implementation.  It is intentionally excluded from routine local
+and CI runs because each partition contains roughly 33B parameters.
+"""
 
 from __future__ import annotations
 
-import hashlib
-import inspect
+import gc
 import os
-from pathlib import Path
 import sys
-from unittest.mock import patch
+from pathlib import Path
+from types import SimpleNamespace
 
 import pytest
 import torch
 from torch.testing import assert_close
 
 
-os.environ["FASTVIDEO_ATTENTION_BACKEND"] = "TORCH_SDPA"
-os.environ["DIFFUSERS_ATTN_BACKEND"] = "native"
-os.environ.setdefault("MASTER_ADDR", "127.0.0.1")
-os.environ.setdefault("MASTER_PORT", "29613")
-os.environ.setdefault("RANK", "0")
-os.environ.setdefault("WORLD_SIZE", "1")
-os.environ.setdefault("LOCAL_RANK", "0")
-
 REPO_ROOT = Path(__file__).resolve().parents[3]
-OFFICIAL_SRC = REPO_ROOT / "DiffusersMiniMaxH3" / "src"
-OFFICIAL_TRANSFORMER = OFFICIAL_SRC / "diffusers" / "models" / "transformers" / "transformer_minimax_h3.py"
-OFFICIAL_TRANSFORMER_SHA256 = "45f1d4a4e9e72128f0d6c5b5d66379b61131a354b58d2776280a83f6c89f8d51"
+OFFICIAL_REF_DIR = Path(os.environ.get("MINIMAX_H3_OFFICIAL_REF_DIR", REPO_ROOT / "DiffusersMiniMaxH3"))
+OFFICIAL_SRC = OFFICIAL_REF_DIR / "src"
+MODEL_ROOT = Path(os.environ.get("MINIMAX_H3_MODEL_ROOT", REPO_ROOT / "official_weights" / "MiniMax-H3"))
+RUN_ENV = "MINIMAX_H3_RUN_DIT_PARITY"
+PARITY_SCOPE = "production_loader"
 
-assert OFFICIAL_TRANSFORMER.is_file(), f"Pinned Diffusers MiniMax H3 source is missing: {OFFICIAL_TRANSFORMER}"
-assert hashlib.sha256(OFFICIAL_TRANSFORMER.read_bytes()).hexdigest() == OFFICIAL_TRANSFORMER_SHA256, (
-    "Diffusers MiniMax H3 reference changed; refresh the pinned source audit before updating parity."
-)
-sys.path.insert(0, str(OFFICIAL_SRC))
-
-from diffusers.models.transformers.transformer_minimax_h3 import (  # noqa: E402
-    MiniMaxH3Transformer3DModel as OfficialMiniMaxH3Transformer,
-)
-
-from fastvideo.configs.models.dits.minimax_h3 import (  # noqa: E402
-    MiniMaxH3ArchConfig,
-    MiniMaxH3Config,
-)
-from fastvideo.distributed import (  # noqa: E402
-    cleanup_dist_env_and_memory,
-    maybe_init_distributed_environment_and_model_parallel,
-)
-from fastvideo.forward_context import set_forward_context  # noqa: E402
-from fastvideo.layers.linear import ReplicatedLinear  # noqa: E402
-from fastvideo.models.dits.minimax_h3 import (  # noqa: E402
-    MiniMaxH3Transformer3DModel as FastVideoMiniMaxH3Transformer,
-)
-from fastvideo.models.loader.fsdp_load import load_model_from_full_model_state_dict  # noqa: E402
-from fastvideo.models.loader.utils import get_param_names_mapping, hf_to_custom_state_dict  # noqa: E402
-from fastvideo.platforms import current_platform  # noqa: E402
-
-
-TINY_CONFIG = {
-    "num_attention_heads": 2,
-    "attention_head_dim": 16,
-    "hidden_size": 24,
-    "num_layers": 2,
-    "num_refiner_layers": 2,
-    "ffn_dim": 32,
-    "in_channels": 4,
-    "audio_in_channels": 6,
-    "patch_size": (1, 2, 2),
-    "text_dim": 8,
-    "freq_dim": 8,
-    "time_embed_hidden_dim": 24,
-    "time_embed_dim": 16,
-    "rope_freq_dim": 2,
-}
-NUM_TEXT_TOKENS = 4
-NUM_AUDIO_TOKENS = 6
-NUM_VIDEO_TOKENS = 8
+if os.environ.get(RUN_ENV) == "1":
+    os.environ.setdefault("FASTVIDEO_ATTENTION_BACKEND", "TORCH_SDPA")
+    os.environ.setdefault("DIFFUSERS_ATTN_BACKEND", "native")
+    os.environ.setdefault("MASTER_ADDR", "127.0.0.1")
+    os.environ.setdefault("MASTER_PORT", "29659")
+    os.environ.setdefault("RANK", "0")
+    os.environ.setdefault("WORLD_SIZE", "1")
+    os.environ.setdefault("LOCAL_RANK", "0")
+    # Put the pinned checkout ahead of the environment's Diffusers installation
+    # before any FastVideo import can populate ``sys.modules``.
+    if OFFICIAL_SRC.is_dir():
+        sys.path.insert(0, str(OFFICIAL_SRC))
 
 
 @pytest.fixture(scope="module", autouse=True)
-def distributed_runtime():
-    """Initialize the SP=1 process groups required by DistributedAttention."""
+def _distributed_runtime():
+    if os.environ.get(RUN_ENV) != "1":
+        yield
+        return
+
+    from fastvideo.distributed import cleanup_dist_env_and_memory, maybe_init_distributed_environment_and_model_parallel
+
     maybe_init_distributed_environment_and_model_parallel(1, 1)
     yield
     cleanup_dist_env_and_memory()
 
 
-def _build_models(param_dtype: torch.dtype = torch.float32) -> tuple[torch.nn.Module, torch.nn.Module]:
-    """Strict-load one deterministic official random state into both models."""
-    torch.manual_seed(20260802)
-    official = OfficialMiniMaxH3Transformer(**TINY_CONFIG).eval()
-    arch_config = MiniMaxH3ArchConfig(**TINY_CONFIG)
-    fastvideo_config = MiniMaxH3Config(arch_config=arch_config)
+def _require_assets(partition: str) -> tuple[torch.device, Path]:
+    if os.environ.get(RUN_ENV) != "1":
+        pytest.skip(f"set {RUN_ENV}=1 on an allocated CUDA node")
+    if not torch.cuda.is_available() or not torch.cuda.is_bf16_supported():
+        pytest.fail("MiniMax H3 DiT parity requires a bf16-capable CUDA GPU", pytrace=False)
+    if not OFFICIAL_SRC.is_dir():
+        pytest.fail(f"pinned Diffusers MiniMax H3 checkout is missing: {OFFICIAL_REF_DIR}", pytrace=False)
 
-    # FastVideo's CPU platform intentionally has no production backend
-    # resolver. This local test selects the existing SDPA implementation only
-    # while constructing DistributedAttention; no production selector changes.
-    with patch.object(
-        current_platform,
-        "get_attn_backend_cls",
-        return_value="fastvideo.attention.backends.sdpa.SDPABackend",
-    ):
-        fastvideo = FastVideoMiniMaxH3Transformer(fastvideo_config, dict(TINY_CONFIG)).eval()
-
-    official_state = official.state_dict()
-    fastvideo_state = fastvideo.state_dict()
-    mapped_state, _ = hf_to_custom_state_dict(
-        official_state,
-        get_param_names_mapping(fastvideo.param_names_mapping),
-    )
-    assert set(mapped_state) == set(fastvideo_state)
-    assert all(fastvideo_state[name].shape == mapped_state[name].shape for name in mapped_state)
-    incompatible = load_model_from_full_model_state_dict(
-        fastvideo,
-        iter(official_state.items()),
-        device=torch.device("cpu"),
-        param_dtype=param_dtype,
-        strict=True,
-        param_names_mapping=get_param_names_mapping(fastvideo.param_names_mapping),
-    )
-    assert not incompatible.missing_keys
-    assert not incompatible.unexpected_keys
-    return official, fastvideo
+    component_dir = MODEL_ROOT / partition
+    required = (component_dir / "config.json", component_dir / "diffusion_pytorch_model.safetensors.index.json")
+    missing = [str(path) for path in required if not path.is_file()]
+    if missing:
+        pytest.fail(f"MiniMax H3 {partition} assets are missing: {missing}", pytrace=False)
+    return torch.device("cuda:0"), component_dir
 
 
 def _make_inputs() -> dict[str, torch.Tensor]:
-    """Build one padless packed layout with three modalities and three times."""
+    """Create a small padless packed layout that exercises all three modalities."""
     generator = torch.Generator(device="cpu").manual_seed(20260803)
-    sequence_length = NUM_TEXT_TOKENS + NUM_AUDIO_TOKENS + NUM_VIDEO_TOKENS
-    text_indices = torch.arange(NUM_TEXT_TOKENS)
-    audio_indices = torch.arange(NUM_TEXT_TOKENS, NUM_TEXT_TOKENS + NUM_AUDIO_TOKENS)
-    video_indices = torch.arange(NUM_TEXT_TOKENS + NUM_AUDIO_TOKENS, sequence_length)
+    num_text_tokens = 2
+    num_audio_tokens = 2
+    num_video_tokens = 4
+    sequence_length = num_text_tokens + num_audio_tokens + num_video_tokens
+
+    text_indices = torch.arange(num_text_tokens)
+    audio_indices = torch.arange(num_text_tokens, num_text_tokens + num_audio_tokens)
+    video_indices = torch.arange(num_text_tokens + num_audio_tokens, sequence_length)
 
     token_tags = torch.empty(sequence_length, dtype=torch.long)
     token_tags[text_indices] = 1
@@ -142,24 +94,13 @@ def _make_inputs() -> dict[str, torch.Tensor]:
 
     position_ids = torch.zeros(sequence_length, 3, dtype=torch.float32)
     position_ids[:, 0] = torch.arange(sequence_length, dtype=torch.float32)
-    position_ids[video_indices, 1] = torch.arange(NUM_VIDEO_TOKENS, dtype=torch.float32) % 4
-    position_ids[video_indices, 2] = torch.arange(NUM_VIDEO_TOKENS, dtype=torch.float32) % 2
+    position_ids[video_indices, 1] = torch.arange(num_video_tokens, dtype=torch.float32) % 2
+    position_ids[video_indices, 2] = torch.arange(num_video_tokens, dtype=torch.float32) // 2
 
-    video_patch_dim = TINY_CONFIG["in_channels"] * 4
     return {
-        "hidden_states": torch.randn(2, NUM_VIDEO_TOKENS, video_patch_dim, generator=generator),
-        "audio_hidden_states": torch.randn(
-            2,
-            NUM_AUDIO_TOKENS,
-            TINY_CONFIG["audio_in_channels"],
-            generator=generator,
-        ),
-        "encoder_hidden_states": torch.randn(
-            2,
-            NUM_TEXT_TOKENS,
-            TINY_CONFIG["text_dim"],
-            generator=generator,
-        ),
+        "hidden_states": torch.randn(1, num_video_tokens, 96, generator=generator),
+        "audio_hidden_states": torch.randn(1, num_audio_tokens, 32, generator=generator),
+        "encoder_hidden_states": torch.randn(1, num_text_tokens, 5120, generator=generator),
         "timestep": torch.tensor([0.7, 0.3, 0.999], dtype=torch.float32),
         "timestep_indices": timestep_indices,
         "token_tags": token_tags,
@@ -170,99 +111,134 @@ def _make_inputs() -> dict[str, torch.Tensor]:
     }
 
 
-def _capture_activations(model: torch.nn.Module, names: tuple[str, ...]) -> tuple[dict[str, list[torch.Tensor]], list]:
-    activations: dict[str, list[torch.Tensor]] = {name: [] for name in names}
-
-    def capture(name: str):
-        def hook(_module, _inputs, output) -> None:
-            if isinstance(output, tuple):
-                output = output[0]
-            assert isinstance(output, torch.Tensor)
-            activations[name].append(output.detach().clone())
-
-        return hook
-
-    modules = dict(model.named_modules())
-    handles = [modules[name].register_forward_hook(capture(name)) for name in names]
-    return activations, handles
+def _to_device(inputs: dict[str, torch.Tensor], device: torch.device) -> dict[str, torch.Tensor]:
+    return {name: tensor.to(device) for name, tensor in inputs.items()}
 
 
-def test_minimax_h3_transformer_matches_pinned_diffusers() -> None:
-    """Compare both output heads numerically with identical weights and rows."""
-    source_file = Path(inspect.getsourcefile(OfficialMiniMaxH3Transformer) or "").resolve()
-    assert source_file == OFFICIAL_TRANSFORMER.resolve()
-    official, fastvideo = _build_models()
-    inputs = _make_inputs()
-    activation_names = (
-        "token_refiner.refiner_blocks.0",
-        "transformer_blocks.0",
-        "norm_out",
-        "proj_out",
-        "audio_proj_out",
+def _assert_mixed_dtype_contract(model: torch.nn.Module) -> None:
+    parameters = dict(model.named_parameters())
+    assert parameters["proj_in.weight"].dtype == torch.float32
+    assert parameters["audio_proj_in.weight"].dtype == torch.float32
+    assert parameters["time_embedder.fc_in.weight" if "time_embedder.fc_in.weight" in parameters
+                      else "time_embedder.linear_1.weight"].dtype == torch.float32
+    assert parameters["proj_out.weight"].dtype == torch.float32
+    assert parameters["audio_proj_out.weight"].dtype == torch.float32
+    assert parameters["context_embedder.weight"].dtype == torch.bfloat16
+    assert parameters["transformer_blocks.0.attn.to_q.weight"].dtype == torch.bfloat16
+
+
+def _load_official(component_dir: Path, device: torch.device) -> torch.nn.Module:
+    from diffusers.models.transformers.transformer_minimax_h3 import MiniMaxH3Transformer3DModel
+    from tests.local_tests.minimax_h3._reference import assert_reference_source
+
+    assert_reference_source(
+        MiniMaxH3Transformer3DModel,
+        "src/diffusers/models/transformers/transformer_minimax_h3.py",
     )
-    official_activations, official_handles = _capture_activations(official, activation_names)
-    fastvideo_activations, fastvideo_handles = _capture_activations(fastvideo, activation_names)
 
+    model, loading_info = MiniMaxH3Transformer3DModel.from_pretrained(
+        component_dir,
+        local_files_only=True,
+        low_cpu_mem_usage=True,
+        torch_dtype=torch.bfloat16,
+        device_map=device,
+        output_loading_info=True,
+    )
+    load_errors = {
+        name: loading_info.get(name, [])
+        for name in ("missing_keys", "unexpected_keys", "mismatched_keys", "error_msgs")
+        if loading_info.get(name)
+    }
+    assert not load_errors, f"official MiniMax H3 checkpoint did not load strictly: {load_errors}"
+    model = model.eval()
+    _assert_mixed_dtype_contract(model)
+    return model
+
+
+def _production_loader_args() -> SimpleNamespace:
+    from fastvideo.configs.pipelines.minimax_h3 import MiniMaxH3PipelineConfig
+
+    return SimpleNamespace(
+        pipeline_config=MiniMaxH3PipelineConfig(),
+        model_paths={},
+        override_transformer_cls_name=None,
+        init_weights_from_safetensors=None,
+        hsdp_replicate_dim=1,
+        hsdp_shard_dim=1,
+        dit_cpu_offload=False,
+        pin_cpu_memory=False,
+        use_fsdp_inference=True,
+        training_mode=False,
+        enable_torch_compile=False,
+        torch_compile_kwargs={},
+        inference_mode=True,
+        dit_layerwise_offload=False,
+    )
+
+
+def _load_fastvideo(component_dir: Path) -> torch.nn.Module:
+    from fastvideo.models.loader.component_loader import TransformerLoader
+
+    model = TransformerLoader().load(str(component_dir), _production_loader_args())
+    _assert_mixed_dtype_contract(model)
+    return model
+
+
+def _run_official(
+    model: torch.nn.Module,
+    inputs: dict[str, torch.Tensor],
+) -> tuple[torch.Tensor, torch.Tensor]:
     with torch.inference_mode():
-        official_video, official_audio = official(**inputs, return_dict=False)
-        with set_forward_context(current_timestep=0, attn_metadata=None):
-            fastvideo_video, fastvideo_audio = fastvideo(**inputs)
-    for handle in official_handles + fastvideo_handles:
-        handle.remove()
-
-    assert_close(fastvideo_video, official_video, atol=1e-5, rtol=1e-5)
-    assert_close(fastvideo_audio, official_audio, atol=1e-5, rtol=1e-5)
-    for name in activation_names:
-        assert len(fastvideo_activations[name]) == len(official_activations[name])
-        for result, expected in zip(fastvideo_activations[name], official_activations[name], strict=True):
-            assert_close(result, expected, atol=1e-5, rtol=1e-5)
+        video, audio = model(**inputs, return_dict=False)
+    return video.detach().float().cpu(), audio.detach().float().cpu()
 
 
-def test_minimax_h3_transformer_preserves_mixed_dtype_islands() -> None:
-    """Keep only the released input/time/output modules in FP32 during loading."""
-    _, fastvideo = _build_models(param_dtype=torch.bfloat16)
+def _run_fastvideo(
+    model: torch.nn.Module,
+    inputs: dict[str, torch.Tensor],
+) -> tuple[torch.Tensor, torch.Tensor]:
+    from fastvideo.forward_context import set_forward_context
 
-    assert fastvideo.proj_in.weight.dtype == torch.float32
-    assert fastvideo.audio_proj_in.weight.dtype == torch.float32
-    assert fastvideo.time_embedder.fc_in.weight.dtype == torch.float32
-    assert fastvideo.proj_out.weight.dtype == torch.float32
-    assert fastvideo.audio_proj_out.weight.dtype == torch.float32
-    assert fastvideo.rope.inv_freq.dtype == torch.float32
-    assert fastvideo.context_embedder.weight.dtype == torch.bfloat16
-    assert fastvideo.token_refiner.refiner_blocks[0].attn.to_q.weight.dtype == torch.bfloat16
-    assert fastvideo.transformer_blocks[0].adaln_proj.linear.weight.dtype == torch.bfloat16
-    assert fastvideo.norm_out.norm.weight.dtype == torch.bfloat16
+    with torch.inference_mode(), set_forward_context(current_timestep=0, attn_metadata=None):
+        video, audio = model(**inputs)
+    return video.detach().float().cpu(), audio.detach().float().cpu()
 
 
-def test_minimax_h3_arch_defaults_match_pinned_reference() -> None:
-    """Pin the released 33B architecture independently of the tiny fixture."""
-    config = MiniMaxH3ArchConfig()
-    assert config.num_layers == 50
-    assert config.num_refiner_layers == 2
-    assert config.hidden_size == 5376
-    assert config.num_attention_heads == 56
-    assert config.attention_head_dim == 128
-    assert config.ffn_dim == 14336
-    assert config.text_dim == 5120
-    assert config.in_channels == 24
-    assert config.audio_in_channels == 32
-    assert 2 * 3 * config.rope_freq_dim == 96
+def _reclaim_vram() -> None:
+    gc.collect()
+    torch.cuda.empty_cache()
 
 
-def test_minimax_h3_transformer_uses_native_projection_surface() -> None:
-    """Keep Diffusers container names at the loader boundary."""
-    _, fastvideo = _build_models()
-    block = fastvideo.transformer_blocks[0]
+def _assert_head_parity(name: str, actual: torch.Tensor, expected: torch.Tensor) -> None:
+    assert actual.shape == expected.shape
+    drift = (actual - expected).abs()
+    reference_abs_mean = expected.abs().mean().item()
+    relative_mean_drift = drift.mean().item() / max(reference_abs_mean, 1e-8)
+    print(
+        f"{name}: reference_abs_mean={reference_abs_mean:.8f} "
+        f"max_abs={drift.max().item():.8f} mean_abs={drift.mean().item():.8f} "
+        f"relative_mean={relative_mean_drift:.8f}",
+        flush=True,
+    )
+    assert relative_mean_drift == 0.0
+    assert_close(actual, expected, atol=0.0, rtol=0.0)
 
-    assert isinstance(fastvideo.time_embedder.fc_in, ReplicatedLinear)
-    assert isinstance(block.attn.to_out, ReplicatedLinear)
-    assert isinstance(block.ff.fc_in, ReplicatedLinear)
-    assert isinstance(block.ff.fc_out, ReplicatedLinear)
-    assert block.ff.fc_in.prefix == "minimax_h3.transformer_blocks.0.ff.fc_in"
-    assert block.attn.to_out.prefix == "minimax_h3.transformer_blocks.0.attn.to_out"
 
-    mapping = get_param_names_mapping(fastvideo.param_names_mapping)
-    assert mapping("time_embedder.linear_1.weight")[0] == "time_embedder.fc_in.weight"
-    assert mapping("transformer_blocks.0.attn.to_out.0.weight")[0] == "transformer_blocks.0.attn.to_out.weight"
-    assert mapping("transformer_blocks.0.ff.net.0.proj.weight")[0] == "transformer_blocks.0.ff.fc_in.weight"
-    assert mapping("transformer_blocks.0.ff.net.2.weight")[0] == "transformer_blocks.0.ff.fc_out.weight"
+@pytest.mark.parametrize("partition", ("transformer", "transformer_ref"))
+def test_minimax_h3_transformer_parity(partition: str) -> None:
+    """Match both DiT partitions against the pinned official implementation."""
+    device, component_dir = _require_assets(partition)
+    cpu_inputs = _make_inputs()
+
+    official = _load_official(component_dir, device)
+    expected_video, expected_audio = _run_official(official, _to_device(cpu_inputs, device))
+    del official
+    _reclaim_vram()
+
+    fastvideo = _load_fastvideo(component_dir)
+    actual_video, actual_audio = _run_fastvideo(fastvideo, _to_device(cpu_inputs, device))
+    del fastvideo
+    _reclaim_vram()
+
+    _assert_head_parity(f"{partition}.video", actual_video, expected_video)
+    _assert_head_parity(f"{partition}.audio", actual_audio, expected_audio)
