+++ tests/local_tests/vaes/test_minimax_h3_audio_vae_parity.py
@@ -1,130 +1,174 @@
 # Copyright 2025 MiniMax authors and HuggingFace Team
 # SPDX-License-Identifier: Apache-2.0
-
-"""CPU synthetic parity for the native MiniMax H3 audio VAE."""
+"""Production-loader parity for the MiniMax H3 audio VAE."""
 
 from __future__ import annotations
 
+import gc
+import os
 import sys
+from pathlib import Path
+from types import SimpleNamespace
 
+import pytest
 import torch
 from torch.testing import assert_close
 
-from tests.local_tests.minimax_h3._reference import REFERENCE_SRC, assert_pinned_reference
 
-assert_pinned_reference(
-    "src/diffusers/models/autoencoders/autoencoder_kl_minimax_h3_audio.py",
-    "1f57b2072c2450ba475e19fc0c1558c394774d9b3ba2291a62f873267d247745",
-)
-sys.path.insert(0, str(REFERENCE_SRC))
+REPO_ROOT = Path(__file__).resolve().parents[3]
+OFFICIAL_REF_DIR = Path(os.environ.get("MINIMAX_H3_OFFICIAL_REF_DIR", REPO_ROOT / "DiffusersMiniMaxH3"))
+OFFICIAL_SRC = OFFICIAL_REF_DIR / "src"
+MODEL_ROOT = Path(os.environ.get("MINIMAX_H3_MODEL_ROOT", REPO_ROOT / "official_weights" / "MiniMax-H3"))
+RUN_ENV = "MINIMAX_H3_RUN_AUDIO_VAE_PARITY"
+PARITY_SCOPE = "production_loader"
 
-from diffusers import AutoencoderKLMiniMaxH3Audio as OfficialMiniMaxH3AudioVAE  # noqa: E402
-
-from fastvideo.configs.models.vaes.minimax_h3_audio import (  # noqa: E402
-    MiniMaxH3AudioVAEArchConfig,
-    MiniMaxH3AudioVAEConfig,
-)
-from fastvideo.models.vaes.minimax_h3_audio import MiniMaxH3AudioVAE  # noqa: E402
-
-
-TINY_ARCH = {
-    "encoder_dim": 4,
-    "encoder_rates": (2, 2),
-    "latent_dim": 32,
-    "latent_channels": 8,
-    "num_attention_heads": 2,
-    "decoder_dim": 16,
-    "decoder_rates": (2, 2),
-    "decoder_kernel_sizes": (4, 4),
-    "resblock_kernel_sizes": (3, 7),
-    "resblock_dilation_sizes": ((1, 3), (1, 3)),
-    "sampling_rate": 32000,
-    "latents_mean": [0.0] * 8,
-    "latents_std": [1.0] * 8,
-}
+if os.environ.get(RUN_ENV) == "1":
+    os.environ.setdefault("MASTER_ADDR", "127.0.0.1")
+    os.environ.setdefault("MASTER_PORT", "29663")
+    os.environ.setdefault("RANK", "0")
+    os.environ.setdefault("WORLD_SIZE", "1")
+    os.environ.setdefault("LOCAL_RANK", "0")
+    if OFFICIAL_SRC.is_dir():
+        sys.path.insert(0, str(OFFICIAL_SRC))
 
 
-def _build_models() -> tuple[OfficialMiniMaxH3AudioVAE, MiniMaxH3AudioVAE]:
-    torch.manual_seed(1234)
-    official = OfficialMiniMaxH3AudioVAE(**TINY_ARCH).float().eval()
+@pytest.fixture(scope="module", autouse=True)
+def _distributed_runtime():
+    if os.environ.get(RUN_ENV) != "1":
+        yield
+        return
 
-    arch = MiniMaxH3AudioVAEArchConfig(**TINY_ARCH)
-    fastvideo = MiniMaxH3AudioVAE(MiniMaxH3AudioVAEConfig(arch_config=arch)).eval()
+    from fastvideo.distributed import cleanup_dist_env_and_memory, maybe_init_distributed_environment_and_model_parallel
 
-    official_state = official.state_dict()
-    fastvideo_state = fastvideo.state_dict()
-    assert official_state.keys() == fastvideo_state.keys()
-    assert {key: value.shape for key, value in official_state.items()} == {
-        key: value.shape for key, value in fastvideo_state.items()
-    }
-    for key in official_state:
-        if key.endswith((".filter", "zero_k_bias")):
-            assert_close(fastvideo_state[key], official_state[key], atol=0.0, rtol=0.0)
-    fastvideo.load_state_dict(official_state, strict=True)
-    return official, fastvideo
+    maybe_init_distributed_environment_and_model_parallel(1, 1)
+    yield
+    cleanup_dist_env_and_memory()
 
 
-def _capture_activations(model: torch.nn.Module, names: tuple[str, ...]) -> tuple[dict[str, list[torch.Tensor]], list]:
-    activations: dict[str, list[torch.Tensor]] = {name: [] for name in names}
+def _require_assets() -> tuple[torch.device, Path]:
+    if os.environ.get(RUN_ENV) != "1":
+        pytest.skip(f"set {RUN_ENV}=1 on an allocated CUDA node")
+    if not torch.cuda.is_available():
+        pytest.fail("MiniMax H3 audio VAE parity requires CUDA", pytrace=False)
+    if not OFFICIAL_SRC.is_dir():
+        pytest.fail(f"pinned Diffusers MiniMax H3 checkout is missing: {OFFICIAL_REF_DIR}", pytrace=False)
 
-    def capture(name: str):
-        def hook(_module, _inputs, output) -> None:
-            assert isinstance(output, torch.Tensor)
-            activations[name].append(output.detach().clone())
-
-        return hook
-
-    modules = dict(model.named_modules())
-    handles = [modules[name].register_forward_hook(capture(name)) for name in names]
-    return activations, handles
+    component_dir = MODEL_ROOT / "audio_vae"
+    required = (component_dir / "config.json", component_dir / "diffusion_pytorch_model.safetensors")
+    missing = [str(path) for path in required if not path.is_file()]
+    if missing:
+        pytest.fail(f"MiniMax H3 audio VAE assets are missing: {missing}", pytrace=False)
+    return torch.device("cuda:0"), component_dir
 
 
-def test_minimax_h3_audio_vae_encode_and_decode_match_reference() -> None:
-    """Compare posterior tensors, direct decode, and the mode round trip."""
+def _load_official(component_dir: Path, device: torch.device) -> torch.nn.Module:
+    from diffusers.models.autoencoders.autoencoder_kl_minimax_h3_audio import AutoencoderKLMiniMaxH3Audio
+    from tests.local_tests.minimax_h3._reference import assert_reference_source
 
-    official, fastvideo = _build_models()
-    generator = torch.Generator(device="cpu").manual_seed(2026)
-    waveform = torch.randn((2, 1, 33), generator=generator, dtype=torch.float32)
-    latents = torch.randn((2, 8, 8), generator=generator, dtype=torch.float32)
-
-    activation_names = ("encoder.block.1", "pre_block", "mean_proj", "decoder.ups.0", "decoder.conv_post")
-    official_activations, official_handles = _capture_activations(official, activation_names)
-    fastvideo_activations, fastvideo_handles = _capture_activations(fastvideo, activation_names)
-    with torch.inference_mode():
-        official_posterior = official.encode(waveform, return_dict=False)[0]
-        fastvideo_posterior = fastvideo.encode(waveform, return_dict=False)[0]
-        official_decode = official.decode(latents).sample
-        fastvideo_decode = fastvideo.decode(latents).sample
-        official_round_trip = official(waveform, sample_posterior=False, return_dict=False)[0]
-        fastvideo_round_trip = fastvideo(waveform, sample_posterior=False, return_dict=False)[0]
-        official_generator = torch.Generator(device="cpu").manual_seed(42)
-        fastvideo_generator = torch.Generator(device="cpu").manual_seed(42)
-        official_sample = official_posterior.sample(generator=official_generator)
-        fastvideo_sample = fastvideo_posterior.sample(generator=fastvideo_generator)
-    for handle in official_handles + fastvideo_handles:
-        handle.remove()
-
-    assert_close(fastvideo_posterior.mode(), official_posterior.mode(), atol=0.0, rtol=0.0)
-    assert_close(fastvideo_posterior.logs, official_posterior.logs, atol=0.0, rtol=0.0)
-    assert_close(fastvideo_sample, official_sample, atol=0.0, rtol=0.0)
-    assert_close(fastvideo_decode, official_decode, atol=0.0, rtol=0.0)
-    assert_close(fastvideo_round_trip, official_round_trip, atol=0.0, rtol=0.0)
-    assert_close(
-        torch.randn(8, generator=fastvideo_generator),
-        torch.randn(8, generator=official_generator),
-        atol=0,
-        rtol=0,
+    assert_reference_source(
+        AutoencoderKLMiniMaxH3Audio,
+        "src/diffusers/models/autoencoders/autoencoder_kl_minimax_h3_audio.py",
     )
 
-    assert fastvideo.hop_length == 4
+    model, loading_info = AutoencoderKLMiniMaxH3Audio.from_pretrained(
+        component_dir,
+        local_files_only=True,
+        low_cpu_mem_usage=True,
+        torch_dtype=torch.float32,
+        device_map=device,
+        output_loading_info=True,
+    )
+    load_errors = {
+        name: loading_info.get(name, [])
+        for name in ("missing_keys", "unexpected_keys", "mismatched_keys", "error_msgs")
+        if loading_info.get(name)
+    }
+    assert not load_errors, f"official MiniMax H3 audio VAE did not load strictly: {load_errors}"
+    model = model.eval()
+    assert all(parameter.dtype == torch.float32 for parameter in model.parameters())
+    return model
+
+
+def _load_fastvideo(component_dir: Path) -> torch.nn.Module:
+    from fastvideo.configs.pipelines.minimax_h3 import MiniMaxH3PipelineConfig
+    from fastvideo.models.loader.component_loader import AudioDecoderLoader
+
+    args = SimpleNamespace(
+        pipeline_config=MiniMaxH3PipelineConfig(),
+        vae_cpu_offload=False,
+    )
+    model = AudioDecoderLoader().load(str(component_dir), args)
+    assert all(parameter.dtype == torch.float32 for parameter in model.parameters())
+    return model
+
+
+def _make_waveform() -> torch.Tensor:
+    generator = torch.Generator(device="cpu").manual_seed(20260803)
+    # An intentionally non-divisible length exercises the production right-pad
+    # path for the model's 800-sample hop length.
+    return torch.randn(1, 1, 1601, generator=generator, dtype=torch.float32)
+
+
+def _run(model: torch.nn.Module, waveform: torch.Tensor) -> dict[str, torch.Tensor]:
+    with torch.inference_mode():
+        posterior = model.encode(waveform, return_dict=False)[0]
+        latents = posterior.mode()
+        decoded = model.decode(latents, return_dict=False)[0]
+    return {
+        "mean": posterior.mean.detach().cpu(),
+        "logs": posterior.logs.detach().cpu(),
+        "decoded": decoded.detach().cpu(),
+    }
+
+
+def _normalize_with_config(model: torch.nn.Module, latents: torch.Tensor) -> torch.Tensor:
+    mean = torch.tensor(model.config.latents_mean, dtype=latents.dtype).view(1, -1, 1)
+    std = torch.tensor(model.config.latents_std, dtype=latents.dtype).view(1, -1, 1)
+    return (latents - mean) / std
+
+
+def _reclaim_vram() -> None:
+    gc.collect()
+    torch.cuda.empty_cache()
+
+
+def _assert_tensor_parity(name: str, actual: torch.Tensor, expected: torch.Tensor, tolerance: float) -> None:
+    assert actual.shape == expected.shape
+    drift = (actual.float() - expected.float()).abs()
+    reference_abs_mean = expected.float().abs().mean().item()
+    relative_mean_drift = drift.mean().item() / max(reference_abs_mean, 1e-8)
+    print(
+        f"{name}: reference_abs_mean={reference_abs_mean:.8f} "
+        f"max_abs={drift.max().item():.8f} mean_abs={drift.mean().item():.8f} "
+        f"relative_mean={relative_mean_drift:.8f}",
+        flush=True,
+    )
+    assert_close(actual, expected, atol=tolerance, rtol=tolerance)
+
+
+def test_minimax_h3_audio_vae_parity() -> None:
+    """Match posterior, normalization, hop geometry, and deterministic decode."""
+    device, component_dir = _require_assets()
+    waveform = _make_waveform()
+
+    official = _load_official(component_dir, device)
+    expected = _run(official, waveform.to(device))
+    expected_normalized = _normalize_with_config(official, expected["mean"])
+    del official
+    _reclaim_vram()
+
+    fastvideo = _load_fastvideo(component_dir)
+    actual = _run(fastvideo, waveform.to(device))
+    actual_normalized = _normalize_with_config(fastvideo, actual["mean"])
+    assert fastvideo.hop_length == 800
     assert fastvideo.sampling_rate == 32000
-    assert fastvideo.latent_channels == 8
-    assert fastvideo.audio_channels == 1
-    assert fastvideo_posterior.mode().shape == (2, 8, 9)
-    assert fastvideo_decode.shape == (2, 1, 32)
-    assert all(parameter.dtype == torch.float32 for parameter in fastvideo.parameters())
-    assert all(buffer.dtype == torch.float32 for buffer in fastvideo.buffers() if buffer.is_floating_point())
-    for name in activation_names:
-        assert len(fastvideo_activations[name]) == len(official_activations[name])
-        for result, expected in zip(fastvideo_activations[name], official_activations[name], strict=True):
-            assert_close(result, expected, atol=1e-6, rtol=1e-6)
+    assert fastvideo.latent_channels == 32
+    del fastvideo
+    _reclaim_vram()
+
+    _assert_tensor_parity("audio_vae.mean", actual["mean"], expected["mean"], 0.0)
+    _assert_tensor_parity("audio_vae.logs", actual["logs"], expected["logs"], 0.0)
+    _assert_tensor_parity("audio_vae.normalized", actual_normalized, expected_normalized, 0.0)
+    # The official and native FP32 BigVGAN paths differ only in the final
+    # weight-norm rounding order (observed max_abs=2.4e-7 on GB200).
+    _assert_tensor_parity("audio_vae.decode", actual["decoded"], expected["decoded"], 5e-7)
